PROFESSIONAL STATEMENT
I am a dedicated and detail-oriented individual seeking an entry-level cybersecurity position. Armed with a solid academic foundation in cybersecurity and hands-on experience gained through internships, I am eager to apply my skills in analyzing and mitigating security risks to contribute to a dynamic team. My passion for staying abreast of evolving cyber threats, coupled with a commitment to continuous learning, positions me as a valuable asset in safeguarding digital assets and ensuring the integrity of organizational information systems.